requirements.txt
python-telegram-bot==13.9
pyyaml==5.3.1
python-binance==1.0.15
